+++ scmintegrations/utils.go
@@ -6,17 +6,15 @@ import (
 	"net/url"
 
 	"github.com/Scalingo/cli/config"
-	"github.com/Scalingo/cli/io"
 	"github.com/Scalingo/go-scalingo/v4"
 	"gopkg.in/errgo.v1"
 )
 
+var ErrNotFound = errors.New("SCM integration not found")
+
 func checkIfIntegrationAlreadyExist(c *scalingo.Client, id string) bool {
 	integrations, _ := c.SCMIntegrationsShow(id)
-	if integrations != nil {
-		return true
-	}
-	return false
+	return integrations != nil
 }
 
 func GetTypeFromURL(integrationURL string) (scalingo.SCMType, error) {
@@ -41,19 +39,5 @@ func GetTypeFromURL(integrationURL string) (scalingo.SCMType, error) {
 		}
 	}
 
-	// If no integration matches the given URL, display a helpful status
-	// message
-	switch u.Host {
-	case "github.com":
-		io.Error("No GitHub integration found, please follow this URL to add it:")
-		io.Errorf("%s/users/github/link\n", config.C.ScalingoAuthUrl)
-	case "gitlab.com":
-		io.Error("No GitLab integration found, please follow this URL to add it:")
-		io.Errorf("%s/users/gitlab/link\n", config.C.ScalingoAuthUrl)
-	default:
-		io.Errorf("No integration found for URL %s.\n", integrationURL)
-		io.Errorf("Please run the command:\n")
-		io.Errorf("scalingo integrations-add gitlab-self-hosted|github-enterprise --url %s://%s --token <personal-access-token>\n", u.Scheme, u.Host)
-	}
-	return "", errors.New("not found")
+	return "", ErrNotFound
 }